+++ src/components/FilterWidget/FilterWidget.styles.ts
@@ -6,6 +6,7 @@ import { colors } from '@keen.io/colors';
 
 export const Container = styled.div<{ isOpen?: boolean }>`
   height: 100%;
+  width: 100%;
   display: flex;
 
   ${(props) =>
@@ -21,13 +22,9 @@ export const TitleContainer = styled.div`
   justify-content: center;
   flex: 1;
   cursor: pointer;
-`;
-
-export const Title = styled.div`
-  font-size: 14px;
-  margin-left: 5px;
-  font-family: 'Lato Bold', sans-serif;
-  color: ${colors.blue[500]};
+  width: 100%;
+  padding: 0 20px;
+  box-sizing: border-box;
 `;
 
 export const DropdownContainer = styled.div<
@@ -95,13 +92,18 @@ export const EmptySearch = styled.div`
 
 export const SelectedPropertiesNumber = styled.div`
   background: ${colors.lightBlue[400]};
-  color: ${colors.black[400]};
   display: inline-flex;
   justify-content: center;
   align-items: center;
   width: 17px;
   height: 17px;
   border-radius: 50%;
-  font-size: 13px;
   margin-left: 2px;
+  flex-shrink: 0;
+`;
+
+export const IconWrapper = styled.div`
+  flex-shrink: 0;
+  display: inline-flex;
+  margin-right: 5px;
 `;