+++ src/modules/chartEditor/saga/openEditor.ts
@@ -1,12 +1,6 @@
 import { take } from 'redux-saga/effects';
-import { scrollToElement } from './scrollToElement';
 import { chartEditorActions } from '../../chartEditor';
 
-export function* openEditor() {
+export function* openEditor(): Generator<unknown, void, boolean> {
   yield take(chartEditorActions.editorMounted.type);
-  const element = document.getElementById('chart-editor');
-
-  if (element) {
-    yield scrollToElement(element);
-  }
 }