+++ src/modules/dashboards/index.ts
@@ -108,11 +108,6 @@ export {
   getDashboardListOrder,
   getCurrentDashboardChartsCount,
   getDashboardAccessKeyRegenerating,
-  DashboardMetaData,
-  DashboardModel,
-  DashboardListOrder,
-  DashboardSettings,
-  DashboardError,
   DASHBOARDS_ORDER,
   ADD_WIDGET_TO_DASHBOARD,
   sortDashboards,
@@ -123,6 +118,14 @@ export {
   setTagsFiltersPublic,
   createDashboardSettings,
   extendDashboardSettings,
+  DashboardError,
   dashboardsSelectors,
   resetDashboardFilters,
+};
+
+export type {
+  DashboardMetaData,
+  DashboardModel,
+  DashboardSettings,
+  DashboardListOrder,
 };