+++ src/modules/datePicker/constants.ts
@@ -3,3 +3,4 @@ export const CLOSE_EDITOR = '@date-picker/CLOSE_EDITOR';
 export const APPLY_EDITOR_SETTINGS = '@date-picker/APPLY_SETTINGS';
 export const SET_EDITOR_CONNECTIONS = '@date-picker/SET_EDITOR_CONNECTIONS';
 export const UPDATE_CONNECTION = '@date-picker/UPDATE_CONNECTION';
+export const SET_NAME = '@date-picker/SET_NAME';