+++ src/modules/datePicker/index.ts
@@ -4,7 +4,9 @@ import {
   applySettings,
   updateConnection,
   setEditorConnections,
+  setName,
 } from './actions';
+
 import datePickerReducer from './reducer';
 import { getDatePickerSettings } from './selectors';
 import { datePickerSaga } from './saga';
@@ -12,6 +14,15 @@ import { datePickerSaga } from './saga';
 import { APPLY_EDITOR_SETTINGS, CLOSE_EDITOR } from './constants';
 
 import { DatePickerConnection, ReducerState } from './types';
+
+const datePickerActions = {
+  openEditor,
+  closeEditor,
+  applySettings,
+  updateConnection,
+  setEditorConnections,
+  setName,
+};
 
 export {
   getDatePickerSettings,
@@ -24,6 +35,7 @@ export {
   datePickerSaga,
   APPLY_EDITOR_SETTINGS,
   CLOSE_EDITOR,
-  ReducerState,
-  DatePickerConnection,
+  datePickerActions,
 };
+
+export type { ReducerState, DatePickerConnection };