+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [3.1.1](https://github.com/keen/explorer/compare/v3.0.3-rc.0...v3.1.1) (2021-07-21)
+
+### Features
+
+* 🎸 Filter widget name added ([052e876](https://github.com/keen/explorer/commit/052e87659c0cafcbd33a7b6dab30a3a6fd62dd3a))
+
+
+### Bug Fixes
+
+* 🐛 CR fixes ([b896e9a](https://github.com/keen/explorer/commit/b896e9a1ed0cf7e53e1c402e8ababab66b42357d))
+* 🐛 CR fixes ([7d1b210](https://github.com/keen/explorer/commit/7d1b2104585bcb881ad076b22b950f71986a78f1))
+* 🐛 dashboard tile animation ([a6c0418](https://github.com/keen/explorer/commit/a6c0418f6116afa8d55400fa68ddabd02a996921))
+* 🐛 Datepicker typography fix ([f1a59c3](https://github.com/keen/explorer/commit/f1a59c3b6c122463703afb444774595bac5e5260))
+* 🐛 Filter actions exported ([5eec2ac](https://github.com/keen/explorer/commit/5eec2aca1337c694cc7b88a1818e45a48b385fbe))
+* 🐛 filter naming ([acf8f3e](https://github.com/keen/explorer/commit/acf8f3ea352b37e10007ef51deb5e042a680354a))
+* 🐛 modal position ([e988397](https://github.com/keen/explorer/commit/e98839781ccee07fae714d34817f2aefc442863a))
+* 🐛 sticky bar overflow ([1a3de6d](https://github.com/keen/explorer/commit/1a3de6deb5b09329331a81965827c1f8ca77d2a0))
+
 ### [3.0.3](https://github.com/keen/explorer/compare/v3.0.3-rc.0...v3.0.3) (2021-07-15)
 
 ### [3.0.2](https://github.com/keen/explorer/compare/v3.0.2-rc.0...v3.0.2) (2021-07-13)